# Conflicts:
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/Models3DController.java
#	src/main/java/pe/edu/pe/smartrent_backend/Controllers/UserController.java
#	src/main/java/pe/edu/pe/smartrent_backend/Repositories/IUserRepository.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesImplements/UserServicesImplements.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IModels3D.java
#	src/main/java/pe/edu/pe/smartrent_backend/ServicesInterfaces/IUser.java
#	src/main/resources/application.properties

+++ src/main/java/pe/edu/pe/smartrent_backend/DTOS/notificationsDTOS/NotificationSecurityAlertDTO.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public class NotificationSecurityAlertDTO {
 
     private LocalDate createdDate;
-    private Long totalAlerts;
+    private int totalAlerts;
 
     public LocalDate getCreatedDate() {
         return createdDate;
@@ -15,11 +15,11 @@ public class NotificationSecurityAlertDTO {
         this.createdDate = createdDate;
     }
 
-    public Long getTotalAlerts() {
+    public int getTotalAlerts() {
         return totalAlerts;
     }
 
-    public void setTotalAlerts(Long totalAlerts) {
+    public void setTotalAlerts(int totalAlerts) {
         this.totalAlerts = totalAlerts;
     }
 }
